Add Error class with passing test
The gem needs to be have its own error class that can parse the
response and fail with so its easier for gem consumers to deal with
errors consistently. I’d add more but we need to make changes
elsewhere. This is just part 1. We should be merging in PR #2 to start
and building off that.