Update test instructions to use .env.sample for faster setup
No point here in linking to dotenv for users to figure out how to set
this up when we have dotenv as a development dependency. Embrace it.
This helps speed up setup and avoid duplication too as we can use this
for development. Also can include more env vars in the future and
maintain it one place.