Fix assertions to use assert_equal to actually compare values
When using `assert` the first argument is supposed to be a boolean, for
example, it could be a logic check. This could have been written as
`assert error_code == 0` but it wasn’t. The second argument for the
`assert` method is the message that appears when it fails so this was a
false positive. We should be using `assert_equal` where the second arg
is a value, not the message.