add oauth_key and expires_in to user model
When interacting with the API we require the oauth_key,
currenlty the authenticate method was not updating this
inside the user model.
I added this because we are transitioning our system to
the new API and we where not using a synapse gem. So this
makes it easier for us to transition to the new version
while not introducing many changes to our system.