Add missing properties to subnet model
As per documentation -
https://docs.synapsefi.com/docs/view-specific-subnet-details - and from
inspecting the raw response, GETing a subnet returns a status field as
well as an account_class field, both of which are important to have and
persist, especially the latter.

I am unable to run the test suite as it seems to depend on auth
environment variables that I do not hav set.